requirements.txt
alembic==1.16.5
blinker==1.9.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
Flask==3.1.2
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
pillow==11.3.0
psycopg2-binary==2.9.11
python-dotenv==1.1.1
reportlab==4.4.4
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.1.3
WTForms==3.1.2
